Eng commentary and name fix
Python module had comments written in my native language, now changed to English. There was an issue, that main packages got names from app_name, that was wrong, so now it was fixed.

+++ Direct/script.sh
@@ -15,10 +15,14 @@ if [ "$char" == "d" ]; then
 	if [ -e $dirname ]; then
 		cd "$dirname"
 
-		rm $curdir/pkg.out
+		if [ -e "$curdir/pkg.out" ]; then
+			rm $curdir/pkg.out
+		fi
 		for filename in ./*.spec
 		do
 			echo "Processing $filename file..."
+			rpmspec -P $filename | grep -e Name >> $curdir/pkg.out
+
 			echo "%mainpackage" >> $curdir/pkg.out
 			# That's our label to show that the new spec file is processed
 			rpmspec -P $filename | grep -e %package -e BuildRequires -e Requires >> $curdir/pkg.out
@@ -37,7 +41,7 @@ elif [ "$char" == "f" ]; then
 		filedir=$(dirname $(realpath "$filename"))
 		cd $filedir
 		filename=$(basename "$filename")
-		rpmspec -P $filename | grep -e %package -e BuildRequires -e Requires > $curdir/pkg.out
+		rpmspec -P $filename | grep -e %package -e Name -e BuildRequires -e Requires > $curdir/pkg.out
 		cd $curdir
 	else
 		echo "Error: $filename does not exist."
